+++ lib/src/flutter_quran_utils.dart
@@ -1002,7 +1002,7 @@ class QuranLibrary {
   /// [hafsStyle] is the default style for Quran so all special characters will be rendered correctly
   final hafsStyle = const TextStyle(
     color: Colors.black,
-    fontSize: 23.55,
+    // fontSize: 23.55,
     fontFamily: "hafs",
     package: "quran_library",
   );